saambaaBidAdapter.js:  make alias of advangelist (#11992)
* saambaaBidAdapter.js: reuse repeated code block

* Update saambaaBidAdapter.js

* Update advangelistsBidAdapter.js

* Update saambaaBidAdapter.js

* Update advangelistsBidAdapter.js

* Update saambaaBidAdapter.js

* Update advangelistsBidAdapter.js

* Create index.js

* Update saambaaBidAdapter.js

* Update advangelistsBidAdapter.js

* Update index.js

* Update advangelistsBidAdapter.js

* Update saambaaBidAdapter.js

* Update index.js

* Update advangelistsBidAdapter.js

* Update saambaaBidAdapter.js

* Update index.js

* Update advangelistsBidAdapter.js

* Update saambaaBidAdapter.js

* Update advangelistsBidAdapter.js

* Delete modules/saambaaBidAdapter.js

* Update advangelistsBidAdapter.js

* Update beachfrontBidAdapter.js

* Update beachfrontBidAdapter.js

* Update beachfrontBidAdapter.js

* Update nextrollBidAdapter.js

* Update beachfrontBidAdapter.js

* Update index.js

* Update advangelistsBidAdapter.js

* Update beachfrontBidAdapter.js

* Create saambaaBidAdapter.js

* fix file name

---------
